--- FirstDemoApp/components/ScrollView.js
@@ -1,43 +1,0 @@
-import React from "react";
-import { StyleSheet, Text, ScrollView, StatusBar, Image } from "react-native";
-// import { SafeAreaView, SafeAreaProvider } from "react-native-safe-area-context";
-
-export default function ScrollViewExample() {
-  return (
-    // <SafeAreaProvider>
-    //   <SafeAreaView style={styles.container} edges={["top"]}>
-    <ScrollView style={styles.scrollView}>
-      <Text style={styles.text}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-        veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-        commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-        velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint
-        occaecat cupidatat non proident, sunt in culpa qui officia deserunt
-        mollit anim id est laborum.
-      </Text>
-      <Image source={require("../assets/splash-icon.png")} />
-    </ScrollView>
-    //   </SafeAreaView>
-    // </SafeAreaProvider>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: StatusBar.currentHeight,
-  },
-  scrollView: {
-    backgroundColor: "pink",
-  },
-  text: {
-    fontSize: 42,
-    padding: 12,
-  },
-  image: {
-    width: 200,
-    height: 200,
-    borderRadius: 10,
-  },
-});
