--- FirstDemoApp/components/helloWorld.js
@@ -1,9 +1,0 @@
-import { View, Text } from "react-native";
-
-export default function HelloWorld() {
-  return (
-    <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-      <Text> Hello, World </Text>
-    </View>
-  );
-}
